+++ lava_scheduler_app/models.py
@@ -55,6 +55,10 @@ class DeviceType(models.Model):
     health_check_job = models.TextField(
         null=True, blank=True, default=None, validators=[validate_job_json])
 
+    @models.permalink
+    def get_absolute_url(self):
+        return ("lava.scheduler.device_type.detail", [self.pk])
+
 
 class Device(models.Model):
     """
@@ -65,12 +69,14 @@ class Device(models.Model):
     IDLE = 1
     RUNNING = 2
     OFFLINING = 3
+    RETIRED = 4
 
     STATUS_CHOICES = (
         (OFFLINE, 'Offline'),
         (IDLE, 'Idle'),
         (RUNNING, 'Running'),
         (OFFLINING, 'Going offline'),
+        (RETIRED, 'Retired'),
     )
 
     # A device health shows a device is ready to test or not